BUG: when initializing to N silks & immediately allocating all N, we sometimes manage to allocate the last silk before it completed processing the BOOT msg. at that time, the queue is already empty but the silk isnt set to FREE state. hence, we crash in silk_alloc() on assertion that alloc must find a FREE silk. we need a generic wait operation of the scheduler bcz more sophisticated schedulers will have more complex logic to indicate they completed their initialization.
Fix:
create an API named silk_eng_is_ready() to check if the engine is ready for use. this will now provide reliable indication of when the inner execution thread has executed all silks to the point they mark their instance as FREE. this prevents the case of a program initializing the Silk library & then failing to allocate some silks bcz not all of them completed booting.