writing a client/server sample application (echo) to show how its possible to implement a network serveru sing Silk employed to service each connection using its own private Silk instance. The code is made to illustrate many ideas of how a server should be written - see the code notes.
The client_echo.c now introduces delay between bytes it sends, thereby causing the server to receive the msg very slow & with many calls to recv(). this showswhat needs to be improved in the echo_server.c to be truly async. in its receive path.
the server is now using a helper function which receives a number of bytes into a buffer & yields between calls to recv() so that the Silk engine thread is always free to do other things. the silk is awakened only when the select() sees it has pending data & sends it a msg. the msg doesnt have information - it only serves to switch the engine back to that Silk.